Main(gord autos)
There are 4 autos in this file. They go from red offensive to defensive to the same for blue.
I probably should have named each one but I forgot which one is which. I know Auto1 corresponds to the 8th slot code. Auto2 corresponds to 7th, Auto3= 6th and Auto4 = 5th slot. 
just change the auto in the autonomous control to whichever file u want to appload